chore(examples): rename examples & fix CSR
