+++ examples/preact/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@lagon/example-preact-ssr-hydration",
+  "name": "@lagon/example-preact",
   "version": "1.0.0",
   "private": true,
   "dependencies": {
