+++ examples/react-ssr/index.tsx
@@ -1,15 +1,24 @@
-import React from 'react';
+import App from './App';
 import { renderToString } from 'react-dom/server';
+import React from 'react';
 
-const App = () => (
-  <div>
-    <h1>Hello World!</h1>
+const html = `<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <meta http-equiv="X-UA-Compatible" content="ie=edge">
+  <link rel="stylesheet" href="/main.css">
+</head>
+<body>
+  <div id="root">
+    ${renderToString(<App />)}
   </div>
-);
+  <script type="module" src="/app.js"></script>
+</body>
+</html>`;
 
-const html = renderToString(<App />);
-
-export async function handler(request: Request): Promise<Response> {
+export function handler(request: Request): Response {
   return new Response(html, {
     headers: {
       'content-type': 'text/html',
