+++ examples/react-ssr/public/main.css
@@ -1,0 +1,7 @@
+body {
+  font-family: Verdana, Geneva, Tahoma, sans-serif;
+}
+
+h1 {
+  color: blue;
+}
